Moved Vulkan Device to Rendering and Changed file name
