+++ IronMaiden/src/Maiden/Rendering/H_Buffer.hpp
@@ -2,7 +2,7 @@
 
 #include "maidenpch.hpp"
 #include "../Core/Main/Core.hpp"
-#include "../Core/H_Device.hpp"
+#include "Vulkan/H_VulkanDevice.hpp"
 
 
 namespace Madam {
