+++ IronMaiden/src/Maiden/Rendering/H_Pipeline.hpp
@@ -2,7 +2,7 @@
 
 #include "maidenpch.hpp"
 #include "../Core/Main/Core.hpp"
-#include "../Core/H_Device.hpp"
+#include "Vulkan/H_VulkanDevice.hpp"
 #include "FrameInfo.hpp"
 
 namespace Madam {
