+++ IronMaiden/src/Maiden/Scene/H_Texture.hpp
@@ -1,11 +1,10 @@
 #pragma once
 
 #include "maidenpch.hpp"
-#include "../Core/H_Device.hpp"
+#include "../Rendering/Vulkan/H_VulkanDevice.hpp"
 
 //libs
 #include <vulkan/vulkan.h>
-
 
 //We might need to move Texture to a struct or have a TextureData struct and Texture builder class??
 namespace Madam {
